qfix

+++ merk/src/estimated_costs/worst_case_costs.rs
@@ -211,7 +211,7 @@ pub fn add_worst_case_merk_propagate(
     cost.storage_cost.replaced_bytes += nodes_updated * MERK_BIGGEST_VALUE_SIZE;
     cost.storage_loaded_bytes +=
         nodes_updated as u64 * (MERK_BIGGEST_VALUE_SIZE + MERK_BIGGEST_KEY_SIZE) as u64;
-    cost.seek_count += nodes_updated as u32;
+    cost.seek_count += nodes_updated;
     cost.hash_node_calls += nodes_updated * 2;
     Ok(())
 }
